# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/fcm-tokens": {
            "get": {
                "description": "Get FCM Token by UserId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FCM Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "userId",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insert FCM Token with UserId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FCM Token"
                ],
                "parameters": [
                    {
                        "description": "JSON data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createTokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete FCM Token with UserId and token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FCM Token"
                ],
                "parameters": [
                    {
                        "description": "JSON data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.deleteTokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
        "/stores/{store_id}/menus": {
            "get": {
                "description": "Get menu by store id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store Menus"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "區分多語系，en為1, zh為2",
                        "name": "language",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用來區分用戶類別，StoreEase商家為0, OrderEase客戶為1",
                        "name": "userType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Menu"
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}/menus": {
            "get": {
                "description": "Get all menus by user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Language",
                        "name": "language",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Menu"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menus"
                ],
                "parameters": [
                    {
                        "description": "JSON data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createMenuInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/menus/{menu_id}": {
            "get": {
                "description": "Get menu menu id with menu owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Language",
                        "name": "language",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Menu"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menus"
                ],
                "parameters": [
                    {
                        "description": "JSON data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.createMenuInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}/stores/{store_id}/menus/{menu_id}": {
            "post": {
                "description": "Create the reference between store and menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Store Id",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu Id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the reference between store and menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Store Id",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu Id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update the reference between store and menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store Menus"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Store Id",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Menu Id",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "isDefault": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.Menu": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "menuItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MenuItem"
                    }
                },
                "store": {
                    "$ref": "#/definitions/domain.Store"
                },
                "title": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "domain.MenuItem": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/domain.Category"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imageBytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "domain.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isBreak": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "storeOpeningHours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.StoreOpeningHour"
                    }
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "domain.StoreOpeningHour": {
            "type": "object",
            "properties": {
                "closeTime": {
                    "type": "string"
                },
                "dayOfWeek": {
                    "type": "integer"
                },
                "openTime": {
                    "type": "string"
                }
            }
        },
...
        "v1.createMenuInput": {
            "type": "object",
            "required": [
                "menuItems",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "menuItems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.menuItemInput"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "v1.createTokenInput": {
            "type": "object",
            "required": [
                "token",
                "userId"
            ],
            "properties": {
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "v1.deleteTokenInput": {
            "type": "object",
            "required": [
                "token",
                "userId"
            ],
            "properties": {
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "v1.menuItemInput": {
            "type": "object",
            "required": [
                "categoryId",
                "price",
                "quantity",
                "title"
            ],
            "properties": {
                "categoryId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "imageBytes": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}